Passes CMake variable LCM_PYTHON_DIR_HINTS to bot2-procman and bot2-lcm-utils
`LCM_PYTHON_DIR_HINTS` allows to find where LCM Python is installed. Without
this variable, the configuration mechanism uses `PYTHON_SITE_PACKAGES_DIR`
which is set in `pods.cmake` and uses the system python site packages directory,
but this may not be correct if LCM has been installed in a different directory,
such as the installation directory of `drake`. In this case, the python site
packages directory used may be different than the one guessed by `pods.cmake`.