Enable creating packages on Ubuntu bionic,xenial, and trusty
Package creation was broken because dpkg-shlibdeps could not find
liblcm.so inside the created package. This is because the package
is installing the tools and the libraries in a non-standard location.
As a work around, we set R(UN)PATH to be able to find liblcm.so
at packaging time. RPATH or RUNPATH are set depending on the
default configuration of the compiler/linker on the current platform.
CMake 3.6 is required to use $ORIGIN in R(UN)PATH for CPackDeb.cmake
module to work.